Begin refactoring resolve-styles into separate files
